from bld-3

+++ resy.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 SOURCES="/workdir/sources"
+SCRIPTS="/workdir/scripts"
 OCI_CONTAINER_X86_64="/workdir/oci-container-x86-64"
 APP_CONTAINER_X86_64="/workdir/app-container-x86-64"
 GITHUB="git://github.com"
@@ -64,7 +65,7 @@ fi
   if [ "$manifest" == "experimental" ]; then
      #export MANIFEST="resy-experimental.xml"
      export META_RESY_BRANCH="dunfell"
-     export META_POKY_BRANCH="2020-04-30-dunfell-3.1"
+     export META_POKY_BRANCH="2020-05-26-dunfell-3.1+"
      export META_VIRTUALIZATION_BRANCH="2020-04-30-dunfell-3.1"
      export META_DESIRE_BRANCH="master"
      export META_MULTI_V7_ML_BSP_BRANCH="dunfell"
@@ -123,40 +124,45 @@ fi
 #repo sync
 
 declare -A MYMAP
-MYMAP[poky]="${GITHUB}/RobertBerger/poky ${SOURCES}/poky ${META_POKY_BRANCH}"
-MYMAP[poky-training]="${GITHUB}/RobertBerger/poky ${SOURCES}/poky-training ${META_POKY_BRANCH}"
+#MYMAP[manifests]="${GITHUB}/RobertBerger/manifests ${SOURCES}/manifests ${MANIFESTS_BRANCH}"
+MYMAP[convenience-scripts]="${GITLAB}/meta-layers/convenience-scripts ${SCRIPTS}/convenience-scripts master"
+#MYMAP[poky]="${GITHUB}/RobertBerger/poky ${SOURCES}/poky ${META_POKY_BRANCH}"
+#MYMAP[poky-training]="${GITHUB}/RobertBerger/poky ${SOURCES}/poky-training ${META_POKY_BRANCH}"
 # my-mender-layer (encrypted)
-MYMAP[my-mender-layer]="${GITHUB}/RobertBerger/my-mender-layer ${SOURCES}/my-mender-layer master"
-MYMAP[meta-virtualization]="${GITHUB}/RobertBerger/meta-virtualization ${SOURCES}/meta-virtualization ${META_VIRTUALIZATION_BRANCH}"
+#MYMAP[my-mender-layer]="${GITHUB}/RobertBerger/my-mender-layer ${SOURCES}/my-mender-layer master"
+#MYMAP[meta-virtualization]="${GITHUB}/RobertBerger/meta-virtualization ${SOURCES}/meta-virtualization ${META_VIRTUALIZATION_BRANCH}"
 # my meta-u-boot-wic-bsp bsp u-boot is here
-MYMAP[meta-u-boot-wic-bsp]="${GITLAB}/meta-layers/meta-u-boot-wic-bsp.git ${SOURCES}/meta-u-boot-wic-bsp ${META_U_BOOT_WIC_BSP_BRANCH}"
+#MYMAP[meta-u-boot-wic-bsp]="${GITLAB}/meta-layers/meta-u-boot-wic-bsp.git ${SOURCES}/meta-u-boot-wic-bsp ${META_U_BOOT_WIC_BSP_BRANCH}"
 # my meta-u-boot-mender-bsp bsp u-boot is here
-MYMAP[meta-u-boot-mender-bsp]="${GITLAB}/meta-layers/meta-u-boot-mender-bsp.git ${SOURCES}/meta-u-boot-mender-bsp ${META_U_BOOT_MENDER_BSP_BRANCH}"
-MYMAP[meta-u-boot-karo-wic-bsp]="${GITLAB}/meta-layers/meta-u-boot-karo-wic-bsp.git ${SOURCES}/meta-u-boot-karo-wic-bsp ${META_U_BOOT_KARO_WIC_BSP_BRANCH}"
-MYMAP[meta-sca]="${GITHUB}/RobertBerger/meta-sca  ${SOURCES}/meta-sca ${META_SCA_BRANCH}"
+#MYMAP[meta-u-boot-mender-bsp]="${GITLAB}/meta-layers/meta-u-boot-mender-bsp.git ${SOURCES}/meta-u-boot-mender-bsp ${META_U_BOOT_MENDER_BSP_BRANCH}"
+#MYMAP[meta-u-boot-karo-wic-bsp]="${GITLAB}/meta-layers/meta-u-boot-karo-wic-bsp.git ${SOURCES}/meta-u-boot-karo-wic-bsp ${META_U_BOOT_KARO_WIC_BSP_BRANCH}"
+#MYMAP[meta-sca]="${GITHUB}/RobertBerger/meta-sca  ${SOURCES}/meta-sca ${META_SCA_BRANCH}"
 # my resy distro 
-MYMAP[resy]="${GITLAB}/meta-layers/meta-resy.git ${SOURCES}/meta-resy ${META_RESY_BRANCH}"
-MYMAP[meta-openembedded]="${GITHUB}/RobertBerger/meta-openembedded ${SOURCES}/meta-openembedded ${META_OPENEMBEDDED_BRANCH}"
-MYMAP[meta-multi-v7-ml-bsp]="${GITLAB}/meta-layers/meta-multi-v7-ml-bsp.git ${SOURCES}/meta-multi-v7-ml-bsp ${META_MULTI_V7_ML_BSP_BRANCH}"
-MYMAP[meta-java]="${GITHUB}/RobertBerger/meta-java ${SOURCES}/meta-java ${META_JAVA_BRANCH}"
-MYMAP[meta-java-examples]="${GITLAB}/meta-layers/meta-java-examples.git ${SOURCES}/meta-java-examples master"
-MYMAP[meta-tensorflow]="${GITHUB}/RobertBerger/meta-tensorflow ${SOURCES}/meta-tensorflow ${META_TENSORFLOW_BRANCH}"
-MYMAP[meta-tensorflow-examples]="${GITLAB}/meta-layers/meta-tensorflow-examples.git ${SOURCES}/meta-tensorflow-examples master"
-MYMAP[meta-golang-examples]="${GITLAB}/meta-layers/meta-golang-examples.git ${SOURCES}meta-golang-examples master"
-MYMAP[manifests]="${GITHUB}/RobertBerger/manifests ${SOURCES}/manifests ${MANIFESTS_BRANCH}"
-MYMAP[meta-python-2]="${GITHUB}/RobertBerger/meta-python2 ${SOURCES}/meta-python2 ${META_PYTHON2_BRANCH}"
+#MYMAP[resy]="${GITLAB}/meta-layers/meta-resy.git ${SOURCES}/meta-resy ${META_RESY_BRANCH}"
+#MYMAP[meta-openembedded]="${GITHUB}/RobertBerger/meta-openembedded ${SOURCES}/meta-openembedded ${META_OPENEMBEDDED_BRANCH}"
+#MYMAP[meta-multi-v7-ml-bsp]="${GITLAB}/meta-layers/meta-multi-v7-ml-bsp.git ${SOURCES}/meta-multi-v7-ml-bsp ${META_MULTI_V7_ML_BSP_BRANCH}"
+#MYMAP[meta-java]="${GITHUB}/RobertBerger/meta-java ${SOURCES}/meta-java ${META_JAVA_BRANCH}"
+#MYMAP[meta-java-examples]="${GITLAB}/meta-layers/meta-java-examples.git ${SOURCES}/meta-java-examples master"
+#MYMAP[meta-tensorflow]="${GITHUB}/RobertBerger/meta-tensorflow ${SOURCES}/meta-tensorflow ${META_TENSORFLOW_BRANCH}"
+#MYMAP[meta-tensorflow-examples]="${GITLAB}/meta-layers/meta-tensorflow-examples.git ${SOURCES}/meta-tensorflow-examples master"
+#MYMAP[meta-golang-examples]="${GITLAB}/meta-layers/meta-golang-examples.git ${SOURCES}/meta-golang-examples master"
+#MYMAP[meta-python-2]="${GITHUB}/RobertBerger/meta-python2 ${SOURCES}/meta-python2 ${META_PYTHON2_BRANCH}"
+#MYMAP[meta-phyboard-polis-imx8mm-bsp]="${GITLAB}/meta-layers/meta-phyboard-polis-imx8mm-bsp ${SOURCES}/meta-phyboard-polis-imx8mm-bsp v4.19.35_1.1.0-phy"
+#MYMAP[meta-bb-syntax]="${GITLAB}/meta-layers/meta-bb-syntax ${SOURCES}/meta-bb-syntax master"
+
+
 
 
 # --> oci
-MYMAP[skopeo-container]="${GITHUB}/RobertBerger/skopeo-container ${OCI_CONTAINER_X86_64}/skopeo-container master"
+#MYMAP[skopeo-container]="${GITHUB}/RobertBerger/skopeo-container ${OCI_CONTAINER_X86_64}/skopeo-container master"
 # <-- oci
 
 # --> app-container-x86-64
-MYMAP[app-container-tensorflow]="${GITLAB}/app-container/app-container-tensorflow.git ${APP_CONTAINER_X86_64}/app-container-tensorflow master"
-MYMAP[app-container-tensorflow-oci]="${GITLAB}/app-container/app-container-tensorflow-oci.git ${APP_CONTAINER_X86_64}/app-container-tensorflow-oci master"
-MYMAP[app-container-java-oci]="${GITLAB}/app-container/app-container-java-oci.git ${APP_CONTAINER_X86_64}/app-container-java-oci master"
-MYMAP[app-container-java-examples-oci]="${GITLAB}/app-container/app-container-java-examples-oci.git ${APP_CONTAINER_X86_64}/app-container-java-examples-oci master"
-MYMAP[app-container-go]="${GITLAB}/app-container/app-container-go.git ${APP_CONTAINER_X86_64}/app-container-go master"
+#MYMAP[app-container-tensorflow]="${GITLAB}/app-container/app-container-tensorflow.git ${APP_CONTAINER_X86_64}/app-container-tensorflow master"
+#MYMAP[app-container-tensorflow-oci]="${GITLAB}/app-container/app-container-tensorflow-oci.git ${APP_CONTAINER_X86_64}/app-container-tensorflow-oci master"
+#MYMAP[app-container-java-oci]="${GITLAB}/app-container/app-container-java-oci.git ${APP_CONTAINER_X86_64}/app-container-java-oci master"
+#MYMAP[app-container-java-examples-oci]="${GITLAB}/app-container/app-container-java-examples-oci.git ${APP_CONTAINER_X86_64}/app-container-java-examples-oci master"
+#MYMAP[app-container-go]="${GITLAB}/app-container/app-container-go.git ${APP_CONTAINER_X86_64}/app-container-go master"
 # <-- app-container-x86-64
 
 #if [ -d ${SOURCES} ]; then
@@ -171,11 +177,78 @@ for K in "${!MYMAP[@]}"
 do
   echo "--> $K"
   set ${MYMAP["${K}"]}
-  # if dir already exists we remove it 
+
+
+  # if dir already exists we need to check if it needs update
   if [ -d $2 ]; then
-     echo "+ rm -rf $2"
-     rm -rf $2 
-  fi
+    pushd $2
+
+    # let's get the latest from remote
+    echo "+ git remote update"
+    git remote update
+
+    # are we on the right branch?
+    LOCAL_BRANCH="$(git rev-parse --abbrev-ref HEAD)"
+    echo "LOCAL_BRANCH: ${LOCAL_BRANCH}"
+
+    BRANCH_WE_WANT="$3"
+    echo "BRANCH_WE_WANT: ${BRANCH_WE_WANT}"
+
+    if [ "$LOCAL_BRANCH" == "$BRANCH_WE_WANT" ]; then
+       echo "branch is already what we want: ${LOCAL_BRANCH}"
+    else
+       echo -e "\e[31mbranch mismatchl\e[39m"
+       echo "+ git checkout $3"
+       git checkout $3
+    fi    
+
+    #echo "+ git remote update"
+    #git remote update
+    #UPSTREAM=${1:-'@{u}'}
+    UPSTREAM="origin/$3"
+    echo "UPSTREAM: ${UPSTREAM}"
+    LOCAL=$(git rev-parse @)
+    echo "LOCAL: ${LOCAL}"
+    REMOTE=$(git rev-parse "$UPSTREAM")
+    echo "REMOTE: ${REMOTE}"
+    BASE=$(git merge-base @ "$UPSTREAM")
+    echo "BASE: ${BASE}"
+
+    if [ $LOCAL = $REMOTE ]; then
+        echo -e "\e[32mUp-to-date\e[39m"
+    elif [ $LOCAL = $BASE ]; then
+        echo -e "\e[31mNeed to pull\e[39m"
+        #echo "+ rm -rf ../$2"
+        echo "+ git pull"
+        echo "Press <ENTER> to go on"
+        read r
+        git pull
+        #rm -rf ../$2
+    elif [ $REMOTE = $BASE ]; then
+        echo "Need to push"
+        echo "Error"
+        echo "Press <ENTER> to go on"
+        read r
+        exit -1
+    else
+        echo "Diverged"
+        echo "Error"
+        echo "Press <ENTER> to go on"
+        read r
+        exit -1
+    fi
+    popd
+ echo "Press <ENTER> to go on"
+ read r
+else # dir exists above
+# dir does not exists here
+# so we need to clone 
+  # if dir already exists we remove it 
+  #if [ -d $2 ]; then
+  #   echo "+ rm -rf $2"
+  #   rm -rf $2 
+  #fi
+
   echo "git clone -b $3 $1 $2"
   git clone -b $3 $1 $2
   retVal=$?
@@ -185,6 +258,8 @@ do
      read r
      exit $retVal
   fi
+fi # dir does not exist
+
   echo "pushd $2"
   pushd $2
   echo "git branch"
@@ -202,14 +277,25 @@ read r
 read r
 
 set -x
-ln -sf sources/manifests/resy-poky-container.sh resy-poky-container.sh
-ln -sf sources/manifests/resy-cooker.sh resy-cooker.sh
-ln -sf sources/manifests/killall_bitbake.sh killall_bitbake.sh
-ln -sf sources/manifests/oci-copy-to-docker.sh oci-copy-to-docker.sh
-ln -sf sources/manifests/build-and-shutdown.sh build-and-shutdown.sh
+if [ ! -L resy-poky-container.sh ]; then
+   ln -sf sources/manifests/resy-poky-container.sh resy-poky-container.sh
+fi
+if [ ! -L resy-cooker.sh ]; then
+   ln -sf sources/manifests/resy-cooker.sh resy-cooker.sh
+fi
+if [ ! -L killall_bitbake.sh ]; then
+   ln -sf sources/manifests/killall_bitbake.sh killall_bitbake.sh
+fi
+if [ ! -L oci-copy-to-docker.sh ]; then
+   ln -sf sources/manifests/oci-copy-to-docker.sh oci-copy-to-docker.sh
+fi
+if [ ! -L build-and-shutdown.sh ]; then
+   ln -sf sources/manifests/build-and-shutdown.sh build-and-shutdown.sh
+fi
 
-
-ln -sf sources/manifests/resy.sh resy.sh
+if [ ! -L resy.sh ]; then
+   ln -sf sources/manifests/resy.sh resy.sh
+fi
 set +x
